# Conflicts:
#	db.sqlite3
#	templates/home.html

+++ PurplePanda/views.py
@@ -45,16 +45,51 @@ class DataView(View):
         return render(request, "viewuser.html", {'print': x, 'current': temp})
 
     def post(self, request):
-        search = request.POST.get('n')
-        try:
-            user = MyUser.objects.get(name=search)
-            MyUser.objects.filter(name=search).delete()
-        except MyUser.DoesNotExist:
-            return HttpResponse("No such user")
-        x = list(MyUser.objects.all())
-        temp = request.session.get("name")
-        temp = MyUser.objects.get(name=temp)
-        return render(request, "viewuser.html", {'print': x, 'current': temp})
+        if "delete" in request.POST:
+            search = request.POST.get('n')
+            try:
+                user = MyUser.objects.get(name=search)
+                MyUser.objects.filter(name=search).delete()
+            except MyUser.DoesNotExist:
+                return HttpResponse("No such user")
+            x = list(MyUser.objects.all())
+            temp = request.session.get("name")
+            temp = MyUser.objects.get(name=temp)
+            return render(request, "viewuser.html", {'print': x, 'current': temp})
+        elif "edit" in request.POST:
+            n = request.POST.get('name')
+            edit = request.POST.get('newname')
+            address = request.POST.get('address')
+            number = request.POST.get('number')
+            p = request.POST.get('password')
+            if n != '':
+                if edit != '':
+                    try:
+                        user2 = MyUser.objects.get(name=n)
+                        try:
+                            user = MyUser.objects.get(name=n)
+                            user.name = edit
+                            if number != '':
+                                user.phoneNumber = number
+                            if address != '':
+                                user.address = address
+                            if p != '':
+                                user.password = p
+                            user.save()
+                            x = list(MyUser.objects.all())
+                            temp = request.session.get("name")
+                            temp = MyUser.objects.get(name=temp)
+                            return render(request, "viewuser.html", {'print': x, 'current': temp})
+                        except MyUser.DoesNotExist:
+                            x = list(MyUser.objects.all())
+                            temp = request.session.get("name")
+                            temp = MyUser.objects.get(name=temp)
+                            return render(request, "viewuser.html", {'print': x, 'current': temp})
+                    except MyUser.DoesNotExist:
+                        x = list(MyUser.objects.all())
+                        temp = request.session.get("name")
+                        temp = MyUser.objects.get(name=temp)
+                        return render(request, "viewuser.html", {'print': x, 'current': temp})
 
 
 class CreateUser(View):
@@ -82,9 +117,19 @@ class CreateUser(View):
 
 class ViewCourses(View):
     def get(self, request):
-        x = list(MyCourses.objects.all())
+        print('viewcourses get')
         temp = request.session.get("name")
         temp = MyUser.objects.get(name=temp)
+
+        if temp.role == "Admin":
+            x = list(MyCourses.objects.all())
+
+        elif temp.role == "TA":
+            x = list(MyCourses.objects.filter(courseTA=temp.name))
+
+        else:
+            x = list(MyCourses.objects.filter(courseInstructor=temp.name))
+
         if x is None:
             return render(request, "viewcourse.html", {"print": "No Courses have been created yet!", "current": temp})
         return render(request, "viewcourse.html", {"print": x, "current": temp})
@@ -97,10 +142,22 @@ class ViewCourses(View):
             course.delete()
         except MyCourses.DoesNotExist:
             return HttpResponse("No such course")
-        x = list(MyCourses.objects.all())
+        temp = request.session.get("name")
+        temp = MyUser.objects.get(name=temp)
+        if temp.role == "Admin" or "admin":
+            x = list(MyCourses.objects.all())
+            print(x)
+        elif temp.role == "ta" or "TA":
+            x = list(MyCourses.objects.all().filter(courseTA=temp.name))
+            print(x)
+
+        else:
+            x = list(MyCourses.objects.all().filter(courseInstructor=temp.name))
+            print(x)
+
         if x is None:
-            return render(request, "viewcourse.html", {"print": "No Courses have been created yet!"})
-        return render(request, "viewcourse.html", {"print": x})
+            return render(request, "viewcourse.html", {"print": "No Courses have been created yet!", "current": temp})
+        return render(request, "viewcourse.html", {"print": x, "current": temp})
 
 
 class CreateCourse(View):
